refactor: shared 폴더에서 공유(전역) 자원 관리
