+++ src/pages/Home/components/sections/BeforeAnswerSection.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 import AnswerTypeSelector from '../AnswerTypeSelector';
-import AnswerButton from '../../../../components/Button/AnswerButton';
+import AnswerButton from '../../../../shared/ui/SharedButton';
 import type { AnswerType } from '../../Home';
 
 interface BeforeAnswerSectionProps {
