+++ src/pages/Subscribe/Subscribe.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import GNB from '../../components/GNB/GNB';
+import GNB from '../../shared/components/Navigation/Navigation';
 import PricingCard from './components/PricingCard';
 import BenefitsCard from './components/BenefitsCard';
