+++ src/pages/Subscribe/components/BenefitsCard.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import { formatCurrency } from '../../../utils/currency';
+import { formatCurrency } from '../../../shared/utils/currency';
 
 interface BenefitsCardProps {
   title: string;
