+++ src/pages/Subscribe/components/PricingCard.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import { formatCurrency } from '../../../utils/currency';
+import { formatCurrency } from '../../../shared/utils/currency';
 
 interface PricingCardProps {
   title: string;
