+++ src/routes/AppRouter.tsx
@@ -4,15 +4,15 @@ import LoginPage from '../pages/Login/Login';
 import ArchivePage from '../pages/Archive/Archive';
 import FeedbackPage from '../pages/Feedback/Feedback';
 import SubscribePage from '../pages/Subscribe/Subscribe';
-import Layout from '../layouts/Layout';
 import HomePage from '../pages/Home/Home';
 import RivalPage from '../pages/Rival/Rival';
 import NotFound from '../pages/NotFound/NotFound';
+import MainLayout from '../shared/layouts/MainLayout';
 
 const AppRouter = () => {
   return (
     <Routes>
-      <Route element={<Layout />}>
+      <Route element={<MainLayout />}>
         <Route path={ROUTE_PATH.LOGIN} element={<LoginPage />} />
         <Route path={ROUTE_PATH.HOME} element={<HomePage />} />
         <Route path={ROUTE_PATH.ARCHIVE} element={<ArchivePage />} />
