+++ src/shared/components/Navigation/TempNav.tsx
@@ -1,6 +1,6 @@
 import styled from '@emotion/styled';
 import { NavLink } from 'react-router-dom';
-import { ROUTE_PATH } from '../routes/routePath';
+import { ROUTE_PATH } from '../../../routes/routePath';
 
 // TODO: 임시 네비게이션 - 정식 GNB 완성 후 삭제 예정
 const TempNav = () => {
